Provided query compilation cache and query cache #65

+++ src/test/java/org/apache/openjpa/azure/TestQueryCompilationCache.java
@@ -18,7 +18,9 @@
  */
 package org.apache.openjpa.azure;
 
+import java.util.List;
 import java.util.Map;
+import org.apache.openjpa.azure.beans.ConfBean;
 import org.apache.openjpa.azure.beans.PObject;
 import org.apache.openjpa.persistence.OpenJPAEntityManager;
 import org.apache.openjpa.persistence.OpenJPAQuery;
@@ -26,19 +28,33 @@ import org.apache.openjpa.persistence.test.SQLListenerTestCase;
 
 public class TestQueryCompilationCache extends SQLListenerTestCase {
 
+    private static boolean initialized = false;
+
     @Override
     public void setUp() {
         super.setUp(
                 "openjpa.QueryCompilationCache", "true",
-                new Class[]{PObject.class}, CLEAR_TABLES);
+                new Class[]{PObject.class, ConfBean.class}, CLEAR_TABLES);
 
-        // check for issue OPENJPA-2301
-        final OpenJPAEntityManager em = emf.createEntityManager();
+        if (!initialized) {
+            // check for issue OPENJPA-2301
+            final OpenJPAEntityManager em = emf.createEntityManager();
 
-        em.getTransaction().begin();
-        em.createQuery("DELETE FROM PObject e");
-        em.getTransaction().commit();
-        em.close();
+            em.getTransaction().begin();
+            em.createQuery("DELETE FROM PObject e");
+            em.createQuery("DELETE FROM ConfBean e");
+
+            ConfBean confBean = new ConfBean();
+            confBean.setKey("chechForQueryResult-key");
+            confBean.setValue("chechForQueryResult-value");
+
+            em.persist(confBean);
+
+            em.getTransaction().commit();
+            em.close();
+
+            initialized = true;
+        }
     }
 
     @Override
@@ -46,7 +62,7 @@ public class TestQueryCompilationCache extends SQLListenerTestCase {
         return System.getProperty("unit", "azure-test");
     }
 
-    public void testQueryCompilationCache()
+    public void testQueryCompilation()
             throws Exception {
 
         OpenJPAEntityManager em = emf.createEntityManager();
@@ -72,5 +88,25 @@ public class TestQueryCompilationCache extends SQLListenerTestCase {
         em.close();
 
         emf.close();
+    }
+
+    public void testForQueryResult() {
+        OpenJPAEntityManager em = emf.createEntityManager();
+
+        em.getTransaction().begin();
+        List<ConfBean> res = em.createQuery("SELECT e FROM ConfBean e").getResultList();
+        assertEquals(1, res.size());
+        assertEquals("chechForQueryResult-value", res.get(0).getValue());
+        em.getTransaction().commit();
+        em.close();
+
+        em = emf.createEntityManager();
+
+        em.getTransaction().begin();
+        res = em.createQuery("SELECT e FROM ConfBean e").getResultList();
+        assertEquals(1, res.size());
+        assertEquals("chechForQueryResult-value", res.get(0).getValue());
+        em.getTransaction().commit();
+        em.close();
     }
 }
